Split index and cli file
